sorting broken into individual components
